requirements.txt
coreapi==2.3.3
Django==4.2.7
django-ckeditor==6.7.0
django-filter==23.4
django-jazzmin==2.6.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
Pillow==10.1.0
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0